Reorganized repo to improve/fix log handling. Corresponds to SDK changes.
