+++ cmd/list.go
@@ -2,11 +2,11 @@ package cmd
 
 import (
 	"fmt"
+	"path"
 
+	hcplugin "github.com/hashicorp/go-plugin"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-
-	"github.com/privateerproj/privateer/run"
 )
 
 var (
@@ -41,12 +41,19 @@ var listCmd = &cobra.Command{
 	},
 }
 
+func init() {
+	rootCmd.AddCommand(listCmd)
+
+	listCmd.PersistentFlags().BoolP("available", "a", false, "Inventory the Armory! List all raids that have been installed.")
+	viper.BindPFlag("available", listCmd.PersistentFlags().Lookup("available"))
+}
+
 // List all available raids and whether or not they're requested in the active config
 // Currently lists raids including the file extension. TODO: should we change that?
 func getAvailableAndRequestedRaids() map[string]string {
 	raids := make(map[string]string)
-	requestedRaids := run.GetRequestedRaids()
-	availableRaids := run.GetAvailableRaids()
+	requestedRaids := GetRequestedRaids()
+	availableRaids := GetAvailableRaids()
 	for _, availableRaid := range availableRaids {
 		raids[availableRaid] = "Not Requested"
 		for _, requestedRaid := range requestedRaids {
@@ -61,8 +68,8 @@ func getAvailableAndRequestedRaids() map[string]string {
 // List only raids requested in the active config and whether or not they're available
 func getRequestedAndAvailableRaids() map[string]string {
 	raids := make(map[string]string)
-	requestedRaids := run.GetRequestedRaids()
-	availableRaids := run.GetAvailableRaids()
+	requestedRaids := GetRequestedRaids()
+	availableRaids := GetAvailableRaids()
 	for _, requestedRaid := range requestedRaids {
 		raids[requestedRaid] = "Not Found"
 		for _, availableRaid := range availableRaids {
@@ -74,9 +81,22 @@ func getRequestedAndAvailableRaids() map[string]string {
 	return raids
 }
 
-func init() {
-	rootCmd.AddCommand(listCmd)
+// GetRequestedRaids returns a list of raid names requested in the config
+func GetRequestedRaids() (raids []string) {
+	if viper.Get("Raids") != nil {
+		raidsVars := viper.Get("Raids").(map[string]interface{})
+		for raidName := range raidsVars {
+			raids = append(raids, raidName)
+		}
+	}
+	return
+}
 
-	listCmd.PersistentFlags().BoolP("available", "a", false, "Inventory the Armory! List all raids that have been installed.")
-	viper.BindPFlag("available", listCmd.PersistentFlags().Lookup("available"))
+// GetAvailableRaids returns a list of raids found in the binaries path
+func GetAvailableRaids() (raids []string) {
+	raidPaths, _ := hcplugin.Discover("*", viper.GetString("binaries-path"))
+	for _, raidPath := range raidPaths {
+		raids = append(raids, path.Base(raidPath))
+	}
+	return
 }
