--- run/approved-raids.go
@@ -1,6 +1,0 @@
-package run
-
-// This is just a single download link for now. Later, stub the URL at the version & infer the filename based on OS type
-var approvedRaids = map[string]string{
-	"wireframe": "https://github.com/privateerproj/raid-hello-world/releases/download/v0.0.0/wireframe",
-}
