--- run/run.go
@@ -1,288 +1,0 @@
-package run
-
-import (
-	"fmt"
-	"os"
-	"os/exec"
-	"os/signal"
-	"path"
-	"path/filepath"
-	"runtime"
-	"strings"
-	"syscall"
-	"io"
-	"net/http"
-	"errors"
-
-	hclog "github.com/hashicorp/go-hclog"
-	hcplugin "github.com/hashicorp/go-plugin"
-	"github.com/spf13/viper"
-
-	"github.com/privateerproj/privateer-sdk/logging"
-	"github.com/privateerproj/privateer-sdk/plugin"
-	"github.com/privateerproj/privateer-sdk/probeengine"
-	"github.com/privateerproj/privateer-sdk/utils"
-)
-
-var logger hclog.Logger
-
-func init() {
-	logger = logging.Logger()
-}
-
-// CLIContext executes all plugins with handling for the command line
-func CLIContext() (err error) {
-	// Setup for handling SIGTERM (Ctrl+C)
-	setupCloseHandler()
-
-	cmdSet, err := getCommands()
-	if err != nil {
-		logger.Error(fmt.Sprintf(
-			"Error loading plugins from config: %s", err))
-		return
-	}
-
-	// Run all plugins
-	err = AllPlugins(cmdSet)
-	if err != nil {
-		logger.Info(fmt.Sprintf(
-			"Output directory: %s", viper.GetString("WriteDirectory")))
-		switch e := err.(type) {
-		case *RaidErrors:
-			logger.Error(fmt.Sprintf(
-				"%d out of %d raids failed.", len(e.Errors), len(cmdSet)))
-			return
-		default:
-			logger.Error(err.Error())
-		}
-	}
-	logger.Info(fmt.Sprintf(
-		"[INFO] No errors encountered during plugin execution. Output directory: %s",
-		viper.GetString("WriteDirectory")))
-	return
-}
-
-// AllPlugins executes specified plugins in a loop
-func AllPlugins(cmdSet []*exec.Cmd) (err error) {
-	// Capture any plugin errors received during execution
-	spErrors := make([]RaidError, 0)
-
-	for _, cmd := range cmdSet {
-		spErrors, err = Plugin(cmd, spErrors)
-		if err != nil {
-			return
-		}
-	}
-
-	if len(spErrors) > 0 {
-		// Return all raid errors to main
-		err = &RaidErrors{
-			Errors: spErrors,
-		}
-	}
-	return
-}
-
-// Plugin executes single plugin based on the provided command
-func Plugin(cmd *exec.Cmd, spErrors []RaidError) ([]RaidError, error) {
-	// Launch the plugin process
-	client := newClient(cmd)
-	defer client.Kill()
-
-	// Connect via RPC
-	rpcClient, err := client.Client()
-	if err != nil {
-		return spErrors, err
-	}
-
-	// Request the plugin
-	rawSP, err := rpcClient.Dispense(plugin.RaidPluginName)
-	if err != nil {
-		return spErrors, err
-	}
-
-	// Execute raid, expecting a silent response
-	raid := rawSP.(plugin.Raid)
-	response := raid.Start()
-	if response != nil {
-		spErr := RaidError{
-			Raid: cmd.String(), // TODO: retrieve raid name from interface function
-			Err:  response,
-		}
-		spErrors = append(spErrors, spErr)
-	} else {
-		logger.Info("Probes all completed with successful results")
-	}
-	return spErrors, nil
-}
-
-// GetRaidBinary returns the path to the executable for the specified raid
-func GetRaidBinary(name string) (binaryName string, err error) {
-	name = filepath.Base(strings.ToLower(name)) // in some cases a filepath may arrive here instead of the base name
-	if runtime.GOOS == "windows" && !strings.HasSuffix(name, ".exe") {
-		name = fmt.Sprintf("%s.exe", name)
-	}
-	plugins, _ := hcplugin.Discover(name, viper.GetString("binaries-path"))
-	if len(plugins) != 1 {
-		err = fmt.Errorf("failed to locate requested plugin '%s' at path '%s'", name, viper.GetString("binaries-path"))
-		return
-	}
-	binaryName = plugins[0]
-
-	return
-}
-
-// setupCloseHandler creates a 'listener' on a new goroutine which will notify the
-// program if it receives an interrupt from the OS. We then handle this by calling
-// our clean up procedure and exiting the program.
-// Ref: https://golangcode.com/handle-ctrl-c-exit-in-terminal/
-func setupCloseHandler() {
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
-	go func() {
-		<-c
-		logger.Error("Execution aborted - SIGTERM")
-		probeengine.CleanupTmp()
-		os.Exit(0)
-	}()
-}
-
-// GetRequestedRaids returns a list of raid names requested in the config
-func GetRequestedRaids() (raids []string) {
-	if viper.Get("Raids") != nil {
-		raidsVars := viper.Get("Raids").(map[string]interface{})
-		for raidName := range raidsVars {
-			raids = append(raids, raidName)
-		}
-	}
-	return
-}
-
-// GetAvailableRaids returns a list of raids found in the binaries path
-func GetAvailableRaids() (raids []string) {
-	raidPaths, _ := hcplugin.Discover("*", viper.GetString("binaries-path"))
-	for _, raidPath := range raidPaths {
-		raids = append(raids, path.Base(raidPath))
-	}
-	return
-}
-
-func getCommands() (cmdSet []*exec.Cmd, err error) {
-	// TODO: give any exec errors a familiar format
-	raids := GetRequestedRaids()
-	for _, raidName := range raids {
-		cmd, err := getCommand(raidName)
-		if err != nil {
-			break
-		}
-		cmdSet = append(cmdSet, cmd)
-	}
-	logger.Debug(fmt.Sprintf(
-		"Using bin: %s", viper.GetString("binaries-path")))
-	if err == nil && len(cmdSet) == 0 {
-		// If there are no errors but also no commands run, it's probably unexpected
-		var available []string
-		GetAvailableRaids()
-		err = utils.ReformatError(
-			"No valid raids specified. Requested: %v, Available: %v", raids, available)
-	}
-	return
-}
-
-func getCommand(raid string) (cmd *exec.Cmd, err error) {
-	binaryName, binErr := GetRaidBinary(raid)
-	if binErr != nil {
-		err = binErr
-		return
-	}
-	cmd = exec.Command(binaryName)
-	flags := fmt.Sprintf("--varsfile=%s", viper.GetString("config")) // TODO this is wonky- can't change it yet without causing an error somewhere else
-	cmd.Args = append(cmd.Args, flags)
-	return
-}
-
-// newClient client handles the lifecycle of a plugin application
-// Plugin hosts should use one Client for each plugin executable
-// (this is different from the client that manages gRPC)
-func newClient(cmd *exec.Cmd) *hcplugin.Client {
-	var pluginMap = map[string]hcplugin.Plugin{
-		plugin.RaidPluginName: &plugin.RaidPlugin{},
-	}
-	var handshakeConfig = plugin.GetHandshakeConfig()
-	return hcplugin.NewClient(&hcplugin.ClientConfig{
-		HandshakeConfig: handshakeConfig,
-		Plugins:         pluginMap,
-		Cmd:             cmd,
-		Logger:          logger,
-	})
-}
-
-// StartApprovedRaid will run a single raid after ensuring it is installed
-// Approved raids are listed in run/approved-raids.go
-func StartApprovedRaid(raidName string) (err error) {
-	configVar := make(map[string]interface{}, 1)
-	configVar[raidName] = make(map[string]interface{}, 1)
-	viper.Set("Raids", configVar)
-
-	err = installIfNotPResent(raidName)
-	if err != nil {
-		logger.Error(fmt.Sprintf(
-			"Installation failed for raid '%s': %v", raidName, err))
-		return
-	}
-	logger.Trace(fmt.Sprintf(
-		"Beginning raid '%s'", raidName))
-	// TODO get the logger set up in the run commands. 
-	// ...Do those benefit from being outside of cmd?
-	err = CLIContext()
-	return
-}
-
-func installIfNotPResent(raidName string) (err error) {
-	installed := false
-	for _, raid := range GetAvailableRaids() {
-		if raid == raidName {
-			installed = true
-		}
-	}
-	if !installed {
-		logger.Trace(fmt.Sprintf(
-			"Installing raid: %s", raidName))
-		err = downloadRaid(raidName)
-	}
-	return err
-}
-
-func downloadRaid(raidName string) (err error) {
-	url := approvedRaids[raidName]
-	if url == "" {
-		return errors.New("raid not found")
-	}
-	logger.Trace(fmt.Sprintf(
-		"Attempting download from: %s", url))
-	resp, err := http.Get(url)
-	if err != nil {
-		return
-	}
-	defer resp.Body.Close()
-
-	u := strings.Split(url, "/")
-	f := u[len(u)-1]
-	localpath := filepath.Join(viper.GetString("binaries-path"), f)
-	logger.Trace(fmt.Sprintf(
-		"Creating file: %s", localpath))
-	out, err := os.Create(localpath)
-	if err != nil {
-		return
-	}
-	defer out.Close()
-
-	logger.Trace("Setting file permissions to 0755")
-	err = out.Chmod(0755)
-	if err != nil {
-		return
-	}
-
-	_, err = io.Copy(out, resp.Body)
-	return
-}
